% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{ldgm_annotation_columns}
\alias{ldgm_annotation_columns}
\title{Extract Selected Annotation Columns}
\usage{
ldgm_annotation_columns(x, ...)
}
\arguments{
\item{x}{Object implementing \link{LdgmAnnotationData}, or an ordinary data frame.}

\item{...}{Reserved for future implementations.}
}
\value{
Character vector of annotation-column names.
}
\description{
S7 generic required by \link{LdgmAnnotationData}.
}
