% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{ldgm_annotation_data_frame}
\alias{ldgm_annotation_data_frame}
\title{Extract an Annotation Data Frame}
\usage{
ldgm_annotation_data_frame(x, ...)
}
\arguments{
\item{x}{Object implementing \link{LdgmAnnotationData}, or an ordinary data frame.}

\item{...}{Reserved for future implementations.}
}
\value{
A data frame containing annotation columns.
}
\description{
S7 generic required by \link{LdgmAnnotationData}.
}
