% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{ldgm_assert_score_test_variant_data}
\alias{ldgm_assert_score_test_variant_data}
\title{Assert Variant Score-Test Row-Data Interface Support}
\usage{
ldgm_assert_score_test_variant_data(x)
}
\arguments{
\item{x}{Data frame or object implementing \link{LdgmScoreTestVariantData}.}
}
\value{
\code{x}, wrapped if needed, invisibly.
}
\description{
Assert Variant Score-Test Row-Data Interface Support
}
