% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{ldgm_assert_summary_stats}
\alias{ldgm_assert_summary_stats}
\title{Assert GraphREML Summary-Statistics Interface Support}
\usage{
ldgm_assert_summary_stats(x)
}
\arguments{
\item{x}{Data frame or object implementing \link{LdgmSummaryStats}.}
}
\value{
\code{x}, wrapped if needed, invisibly.
}
\description{
Assert GraphREML Summary-Statistics Interface Support
}
