% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{ldgm_block_directory}
\alias{ldgm_block_directory}
\title{Extract an LDGM Block Directory}
\usage{
ldgm_block_directory(x, ...)
}
\arguments{
\item{x}{Object implementing \link{LdgmBlockCatalog}, a metadata data frame, or a
metadata CSV path.}

\item{...}{Reserved for future implementations.}
}
\value{
Single directory path containing metadata-referenced block files.
}
\description{
S7 generic required by \link{LdgmBlockCatalog}.
}
