% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{ldgm_block_metadata_frame}
\alias{ldgm_block_metadata_frame}
\title{Extract an LDGM Block Metadata Frame}
\usage{
ldgm_block_metadata_frame(x, ...)
}
\arguments{
\item{x}{Object implementing \link{LdgmBlockCatalog}, a metadata data frame, or a
metadata CSV path.}

\item{...}{Reserved for future implementations.}
}
\value{
A GraphLD-style metadata data frame.
}
\description{
S7 generic required by \link{LdgmBlockCatalog}.
}
