% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ldgm_precision_matrix}
\alias{ldgm_precision_matrix}
\title{Extract an LDGM Precision Matrix}
\usage{
ldgm_precision_matrix(x)
}
\arguments{
\item{x}{An \code{ldgm_precision} object or sparse matrix.}
}
\value{
A \code{Matrix::dgCMatrix} precision matrix.
}
\description{
Extract an LDGM Precision Matrix
}
