% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ldgm_read_snplist}
\alias{ldgm_read_snplist}
\title{Read an LDGM SNP List File}
\usage{
ldgm_read_snplist(path)
}
\arguments{
\item{path}{Path to the snplist file.}
}
\value{
A data frame.
}
\description{
Reads a comma-separated \code{.snplist} file as a data frame while preserving column
names used by \code{ldgm} and GraphLD.
}
