% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{ldgm_score_test_gene_data}
\alias{ldgm_score_test_gene_data}
\title{Wrap Gene Score-Test Row Data}
\usage{
ldgm_score_test_gene_data(x)
}
\arguments{
\item{x}{Data frame, or an existing object implementing
\link{LdgmScoreTestGeneData}.}
}
\value{
An object implementing \link{LdgmScoreTestGeneData}.
}
\description{
Converts an ordinary R data frame into the default S7 implementation of the
\link{LdgmScoreTestGeneData} interface.
}
