% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{ldgm_score_test_variant_data}
\alias{ldgm_score_test_variant_data}
\title{Wrap Variant Score-Test Row Data}
\usage{
ldgm_score_test_variant_data(x)
}
\arguments{
\item{x}{Data frame, or an existing object implementing
\link{LdgmScoreTestVariantData}.}
}
\value{
An object implementing \link{LdgmScoreTestVariantData}.
}
\description{
Converts an ordinary R data frame into the default S7 implementation of the
\link{LdgmScoreTestVariantData} interface.
}
