% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{ldgm_summary_stats_frame}
\alias{ldgm_summary_stats_frame}
\title{Extract a Summary-Statistics Data Frame}
\usage{
ldgm_summary_stats_frame(x, ...)
}
\arguments{
\item{x}{Object implementing \link{LdgmSummaryStats}, or an ordinary data frame.}

\item{...}{Reserved for future implementations.}
}
\value{
A data frame with at least \code{SNP} and \code{Z} by default.
}
\description{
S7 generic required by \link{LdgmSummaryStats}.
}
