% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ldgm_variant_info}
\alias{ldgm_variant_info}
\title{Extract LDGM Variant Information}
\usage{
ldgm_variant_info(x)
}
\arguments{
\item{x}{An \code{ldgm_precision} object.}
}
\value{
A data frame of variant metadata.
}
\description{
Extract LDGM Variant Information
}
