% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\docType{data}
\name{LdgmAnnotationData}
\alias{LdgmAnnotationData}
\title{GraphREML Annotation-Data Interface}
\format{
An object of class \code{s7contract::s7_interface} (inherits from \code{S7_object}) of length 1.
}
\usage{
LdgmAnnotationData
}
\value{
An \code{s7contract} interface object.
}
\description{
Structural interface for upstream GraphLD-style \code{annotation_data} inputs. In
Python GraphLD this argument is a \code{polars.DataFrame}; in RcppLDGM it is any
object that implements \code{\link[=ldgm_annotation_data_frame]{ldgm_annotation_data_frame()}} and
\code{\link[=ldgm_annotation_columns]{ldgm_annotation_columns()}}. Use \code{\link[=ldgm_annotation_data]{ldgm_annotation_data()}} to wrap ordinary
data frames in the default S7 implementation.
}
\keyword{datasets}
