% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\docType{data}
\name{LdgmScoreTestVariantData}
\alias{LdgmScoreTestVariantData}
\title{GraphLD Variant Score-Test Row-Data Interface}
\format{
An object of class \code{s7contract::s7_interface} (inherits from \code{S7_object}) of length 1.
}
\usage{
LdgmScoreTestVariantData
}
\value{
An \code{s7contract} interface object.
}
\description{
Structural interface for variant-level row-data inputs written to GraphLD
score-test HDF5 files. In Python GraphLD this is a table-like object with at
least \code{CHR}, \code{POS}, and \code{RSID}/\code{SNP}. In RcppLDGM it is any object that
implements \code{\link[=ldgm_score_test_variant_data_frame]{ldgm_score_test_variant_data_frame()}}. Use
\code{\link[=ldgm_score_test_variant_data]{ldgm_score_test_variant_data()}} to wrap ordinary data frames.
}
\keyword{datasets}
