% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\docType{data}
\name{LdgmSummaryStats}
\alias{LdgmSummaryStats}
\title{GraphREML Summary-Statistics Interface}
\format{
An object of class \code{s7contract::s7_interface} (inherits from \code{S7_object}) of length 1.
}
\usage{
LdgmSummaryStats
}
\value{
An \code{s7contract} interface object.
}
\description{
Structural interface for upstream GraphLD-style \code{summary_stats} inputs. In
Python GraphLD this argument is a \code{polars.DataFrame}; in RcppLDGM it is any
object that implements \code{\link[=ldgm_summary_stats_frame]{ldgm_summary_stats_frame()}}. Use
\code{\link[=ldgm_summary_stats]{ldgm_summary_stats()}} to wrap ordinary data frames in the default S7
implementation.
}
\keyword{datasets}
