% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{RcppLDGM-package}
\alias{RcppLDGM}
\alias{RcppLDGM-package}
\title{RcppLDGM: Rcpp Port of Linkage Disequilibrium Graphical Models}
\description{
RcppLDGM is an incremental, compatibility-tested Rcpp port of the upstream
Python \code{ldgm} package. The first exported functions expose native graph
primitives used by the LDGM reduction algorithm; tree-sequence bricking and
full LDGM construction will be added behind the same compatibility contract.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sounkou-bioinfo/RCppLDGM}
  \item Report bugs at \url{https://github.com/sounkou-bioinfo/RCppLDGM/issues}
}

}
\author{
\strong{Maintainer}: Sounkou Toure \email{sounkoutoure@gmail.com}

Other contributors:
\itemize{
  \item Anthony Wilder Wohns [copyright holder]
  \item Luke O'Connor [copyright holder]
  \item Pouria Salehi Nowbandegani [copyright holder]
  \item RcppLDGM contributors [contributor]
}

}
\keyword{internal}
